rearrange APIs for better guidance
